Move RightIcons to Components and create index
