+++ app/containers/message/Components/RightIcons/MessageError.tsx
@@ -1,12 +1,12 @@
 import React, { useContext } from 'react';
 
-import Touchable from '../Touchable';
-import { CustomIcon } from '../../CustomIcon';
-import styles from '../styles';
-import { BUTTON_HIT_SLOP } from '../utils';
-import { themes } from '../../../lib/constants';
-import MessageContext from '../Context';
-import { useTheme } from '../../../theme';
+import Touchable from '../../Touchable';
+import { CustomIcon } from '../../../CustomIcon';
+import styles from '../../styles';
+import { BUTTON_HIT_SLOP } from '../../utils';
+import { themes } from '../../../../lib/constants';
+import MessageContext from '../../Context';
+import { useTheme } from '../../../../theme';
 
 const MessageError = React.memo(
 	({ hasError }: { hasError: boolean }) => {
