+++ app/containers/message/Message.tsx
@@ -19,7 +19,7 @@ import CallButton from './CallButton';
 import { themes } from '../../lib/constants';
 import { IMessage, IMessageInner, IMessageTouchable } from './interfaces';
 import { useTheme } from '../../theme';
-import RightIcons from './Icons/RightIcons';
+import RightIcons from './Components/RightIcons';
 
 const MessageInner = React.memo((props: IMessageInner) => {
 	const { attachments } = props;
